ref: move shared/typings to types/api
